refactor(murmur): split ui.rs into one submodule per frame region
Pure movement, no behaviour change: `ui.rs` was 2,668 lines against the
800-line rule. It is now the layout root (`render`, the overlay and the
slash-command popup) plus six submodules under `ui/`, one per region of
the frame — `band` (live transcript + flush to scrollback; its tests in
`band/tests.rs`, same shape as `fleet_rail/tests.rs`), `message`,
`chooser`, `rail`, `status`, `hitl`. Largest file is now 517 lines.

Mechanical edits only: moved items are `pub(super)` (`hitl_scroll_step`
`pub(crate)` so the root can re-export it to `mod.rs`), `super::x` paths
inside the moved code point at `crate::cmd::agent::cli::x`, and each
file imports exactly what it uses. Audited by normalising visibility and
paths on both sides and diffing sorted lines: the only differences are
rustfmt reflow of three signatures, the import lines, and the re-export.
