+++ mur-core/src/cmd/agent/cli/ui.rs
@@ -1,20 +1,34 @@
 //! Ratatui rendering: transcript pane, input box, status bar, HITL modal.
+//!
+//! One submodule per region of the frame: `band` (the live transcript and its
+//! flush to scrollback), `message` (one message's lines), `chooser`, `rail`,
+//! `status`, `hitl`. This root owns the layout and the two popups that sit on
+//! the composer.
+
+mod band;
+mod chooser;
+mod hitl;
+mod message;
+mod rail;
+mod status;
+
+pub use band::flush_finished;
+pub(super) use hitl::hitl_scroll_step;
+
+use band::render_transcript;
+use chooser::{chooser_band_height, render_chooser_band};
+use hitl::render_hitl;
+use rail::{fleet_rail_height, render_fleet_rail};
+use status::render_status;
 
 use ratatui::Frame;
-use ratatui::backend::Backend;
 use ratatui::layout::{Constraint, Direction, Layout, Rect};
 use ratatui::style::{Color, Modifier, Style};
 use ratatui::text::{Line, Span, Text};
-use ratatui::widgets::{
-    Block, Borders, Clear, List, ListItem, ListState, Padding, Paragraph, Widget, Wrap,
-};
+use ratatui::widgets::{Block, Borders, Clear, List, ListItem, ListState, Padding, Paragraph};
 
-use std::time::Instant;
-
-use super::app::{App, ChatMsg, Role, SPINNER, Severity};
+use super::app::App;
 use super::complete;
-use super::markdown;
-use super::welcome::welcome_lines;
 
 /// Footer hint shown at the bottom of the full-screen transcript overlay
 /// (Ctrl+O). Enter and Esc both return to chat; Ctrl+D quits outright, same
@@ -22,50 +36,11 @@ use super::welcome::welcome_lines;
 /// unhandled into the input box.
 const OVERLAY_HINT: &str = " press Enter or Esc to return · Ctrl+D quit ";
 
-/// Body indent under a role header ("you ›" / "● agent") so a message's content
-/// reads as belonging to its speaker rather than sitting flush with the header.
-const MSG_INDENT: &str = "  ";
-
 /// Composer height when the input is empty (one text row plus its border).
 /// Typing grows it up to `INPUT_H_MAX`.
-const INPUT_H_MIN: u16 = 3;
+pub(super) const INPUT_H_MIN: u16 = 3;
+
 const INPUT_H_MAX: u16 = 8;
-
-/// Approval modal size, as a percentage of the viewport.
-const HITL_PCT_X: u16 = 70;
-const HITL_PCT_Y: u16 = 50;
-
-/// Rows one PgUp/PgDn moves the approval modal's body, used only until the
-/// modal has been drawn once and can report its real height.
-///
-/// A fixed step is not safe on its own: when the box is short (a narrow pane
-/// wraps the input into more rows, leaving as few as two visible) a step of 5
-/// jumps clean over the rows in between, and the operator is never shown them.
-/// Skipping content in the one modal whose whole job is "read this before it
-/// runs" is the worst place to lose a line, so `hitl_scroll_step` clamps the
-/// step to what the renderer last had room for.
-pub(super) const HITL_SCROLL_PAGE: u16 = 5;
-
-/// Rows to page by, given how many body rows the modal last displayed.
-///
-/// Never larger than the visible window, so paging cannot step over a row that
-/// was never on screen. Keeps one row of overlap for reading continuity, and
-/// falls back to [`HITL_SCROLL_PAGE`] before the first draw reports a height.
-pub(super) const fn hitl_scroll_step(visible_rows: u16) -> u16 {
-    if visible_rows == 0 {
-        HITL_SCROLL_PAGE
-    } else if visible_rows > 1 {
-        visible_rows - 1
-    } else {
-        1
-    }
-}
-
-/// Prepend the body indent to an already-styled line (e.g. cached markdown).
-fn indent_line(mut line: Line<'static>) -> Line<'static> {
-    line.spans.insert(0, Span::raw(MSG_INDENT));
-    line
-}
 
 /// Draw the whole UI for one frame.
 pub fn render(f: &mut Frame, app: &mut App) {
@@ -283,2386 +258,4 @@ fn render_completion(f: &mut Frame, app: &App, input_area: Rect) {
         popup_area,
         &mut list_state,
     );
-}
-
-/// Height of the chooser band, or 0 when no agent chooser is open (slash
-/// menu stays a popup). Full spaced rows (label + optional desc + spacer)
-/// when they fit above the composer while leaving the transcript at least
-/// `MIN_TRANSCRIPT_ROWS`; otherwise compact one-line rows; never taller
-/// than the space available (the List scrolls the selection into view).
-const MIN_TRANSCRIPT_ROWS: u16 = 3;
-
-/// Share of the viewport the transcript keeps when the chooser is open,
-/// in percent.
-const TRANSCRIPT_FLOOR_PCT: u16 = 40;
-
-fn chooser_band_height(app: &App, total_h: u16, input_height: u16) -> u16 {
-    let Some(state) = &app.completion else {
-        return 0;
-    };
-    if !state.spaced || state.items.is_empty() {
-        return 0;
-    }
-    let chrome = input_height + 1; // composer + status line
-    let full: u16 = state
-        .items
-        .iter()
-        .map(|c| 2 + u16::from(!c.desc.is_empty())) // label + spacer (+ desc)
-        .sum::<u16>()
-        .saturating_add(2); // borders
-    let compact = (state.items.len() as u16).saturating_add(2);
-    // Prefer the readable floor; fall back to the hard minimum only when the
-    // floor would squeeze the chooser below its compact form. The chooser is
-    // what the operator must act on, so it never loses this trade.
-    let roomy = total_h
-        .saturating_sub(chrome)
-        .saturating_sub((total_h * TRANSCRIPT_FLOOR_PCT / 100).max(MIN_TRANSCRIPT_ROWS));
-    let tight = total_h.saturating_sub(chrome + MIN_TRANSCRIPT_ROWS);
-    let available = if roomy >= compact { roomy } else { tight };
-    // Take `compact` exactly when the spaced form does not fit — padding the
-    // band out to `available` spends rows on nothing.
-    let auto = if full <= available {
-        full
-    } else {
-        compact.min(available).max(3)
-    };
-    // Ctrl+↑/↓ while the chooser is open grows/shrinks the band on top of
-    // the auto height, clamped so the transcript keeps its minimum rows.
-    (i32::from(auto) + i32::from(app.chooser_grow))
-        .clamp(3, i32::from(tight.max(MIN_TRANSCRIPT_ROWS))) as u16
-}
-
-/// Draw the agent chooser into its own layout band. Falls back to compact
-/// one-line rows ("N label — desc") when the band is shorter than the full
-/// spaced form.
-fn render_chooser_band(f: &mut Frame, app: &App, area: Rect) {
-    let Some(state) = &app.completion else {
-        return;
-    };
-    let theme = app.theme;
-    let full: u16 = state
-        .items
-        .iter()
-        .map(|c| 2 + u16::from(!c.desc.is_empty()))
-        .sum::<u16>()
-        .saturating_add(2);
-    let compact = area.height < full;
-
-    let rows: Vec<ListItem> = state
-        .items
-        .iter()
-        .enumerate()
-        .map(|(i, c)| {
-            if compact {
-                let mut spans = vec![
-                    Span::styled(format!("{} ", i + 1), Style::default().fg(theme.system)),
-                    Span::styled(c.display.clone(), Style::default().fg(theme.agent_text)),
-                ];
-                if !c.desc.is_empty() {
-                    spans.push(Span::styled(
-                        format!(" — {}", c.desc),
-                        Style::default().fg(Color::DarkGray),
-                    ));
-                }
-                ListItem::new(Line::from(spans))
-            } else {
-                let mut lines = vec![Line::from(vec![
-                    Span::styled(format!("{} ", i + 1), Style::default().fg(theme.system)),
-                    Span::styled(c.display.clone(), Style::default().fg(theme.agent_text)),
-                ])];
-                if !c.desc.is_empty() {
-                    lines.push(Line::from(Span::styled(
-                        format!("   {}", c.desc),
-                        Style::default().fg(Color::DarkGray),
-                    )));
-                }
-                lines.push(Line::default());
-                ListItem::new(lines)
-            }
-        })
-        .collect();
-
-    let block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(Style::default().fg(theme.border))
-        .padding(Padding::horizontal(1))
-        .title(" 1-9 pick · ↑↓ move · Enter accept · Esc close · Ctrl+↑↓ resize ")
-        .title_style(Style::default().fg(theme.border_title));
-
-    let mut list_state = ListState::default();
-    list_state.select(Some(state.selected));
-
-    f.render_stateful_widget(
-        List::new(rows)
-            .block(block)
-            .highlight_style(
-                Style::default()
-                    .fg(theme.accent)
-                    .add_modifier(Modifier::BOLD),
-            )
-            .highlight_symbol("❯ "),
-        area,
-        &mut list_state,
-    );
-}
-
-/// Fixed separator width for flushed scrollback lines: the real frame width
-/// isn't known at flush time (content prints above the inline viewport and
-/// the terminal soft-wraps it), so a modest fixed rule stands in.
-const SEPARATOR_WIDTH: usize = 60;
-
-/// Rows the fleet rail paints: one collapsed line, plus a capped member list
-/// when someone is blocked, plus one more row for the "… N more" truncation
-/// notice when the member list doesn't fit. Must always equal
-/// `rail_lines(view, _).len()` — see `the_rail_height_matches_painted_lines`.
-/// A working fleet is not news; a stalled one is.
-pub fn rail_height_for(view: &crate::cmd::agent::cli::fleet_rail::RailView) -> u16 {
-    use crate::cmd::agent::cli::fleet_rail::{MAX_EXPANDED_ROWS, MemberState};
-    let blocked = view
-        .members
-        .iter()
-        .any(|m| matches!(m.state, MemberState::Blocked { .. }));
-    if !blocked {
-        return 1;
-    }
-    let shown = view.members.len().min(MAX_EXPANDED_ROWS) as u16;
-    let truncated = view.members.len() > MAX_EXPANDED_ROWS;
-    1 + shown + u16::from(truncated)
-}
-
-/// Height of the rail band for the current app state; 0 when `--fleet` is off.
-pub fn fleet_rail_height(app: &App) -> u16 {
-    app.fleet_view().map(rail_height_for).unwrap_or(0)
-}
-
-/// The fleet rail's content as plain lines: the head line, a capped member
-/// list, and a truncation notice when the list doesn't fit. Pulled out of
-/// `render_fleet_rail` so `rail_height_for` has something concrete to be
-/// tested against instead of a number nothing checks — the truncation notice
-/// is the thing most likely to silently fall off the bottom again.
-fn rail_lines(
-    view: &crate::cmd::agent::cli::fleet_rail::RailView,
-    theme: &'static super::theme::Theme,
-) -> Vec<Line<'static>> {
-    use crate::cmd::agent::cli::fleet_rail::{MAX_EXPANDED_ROWS, MemberState};
-    let mut lines: Vec<Line> = Vec::new();
-
-    let head = match &view.notice {
-        Some(n) => format!("{}  {n}", view.jobs_line),
-        None => view.jobs_line.clone(),
-    };
-    lines.push(Line::styled(
-        head,
-        Style::default()
-            .fg(theme.border_title)
-            .add_modifier(Modifier::BOLD),
-    ));
-
-    if rail_height_for(view) > 1 {
-        for m in view.members.iter().take(MAX_EXPANDED_ROWS) {
-            let (body, color) = match &m.state {
-                MemberState::Blocked { summary, .. } => (format!("blocked: {summary}"), theme.warn),
-                MemberState::Working { tool, since } => (
-                    match tool {
-                        Some(t) => format!("working ({}) · {t}", elapsed(*since)),
-                        None => format!("working ({})", elapsed(*since)),
-                    },
-                    theme.agent,
-                ),
-                MemberState::Done => ("done".to_string(), theme.success),
-                MemberState::Failed => ("failed".to_string(), theme.error),
-            };
-            let glyph = m.state.glyph();
-            lines.push(Line::styled(
-                format!("  {:<10} {glyph} {body}", m.agent),
-                Style::default().fg(color),
-            ));
-        }
-        let extra = view.members.len().saturating_sub(MAX_EXPANDED_ROWS);
-        if extra > 0 {
-            lines.push(Line::styled(
-                format!("  … {extra} more"),
-                Style::default().fg(theme.system),
-            ));
-        }
-    }
-
-    lines
-}
-
-fn render_fleet_rail(f: &mut Frame, app: &App, area: Rect) {
-    let Some(view) = app.fleet_view() else {
-        return;
-    };
-    f.render_widget(
-        Paragraph::new(Text::from(rail_lines(view, app.theme))),
-        area,
-    );
-}
-
-/// "2m" / "1h04m" — elapsed since a member last changed state. Shown instead
-/// of a staleness verdict: a runtime that died mid-turn shows a growing
-/// number rather than a state we guessed.
-fn elapsed(since: chrono::DateTime<chrono::Utc>) -> String {
-    let secs = (chrono::Utc::now() - since).num_seconds().max(0);
-    match secs {
-        0..=59 => format!("{secs}s"),
-        60..=3599 => format!("{}m", secs / 60),
-        _ => format!("{}h{:02}m", secs / 3600, (secs % 3600) / 60),
-    }
-}
-
-/// Rows left for the live transcript band inside `viewport_h` once the
-/// composer, status line, chooser band and fleet rail have taken theirs. The
-/// band draws no border of its own (see `render_transcript`), so this is the
-/// exact row count it paints into.
-fn band_capacity(viewport_h: u16, input_h: u16, chooser_h: u16, rail_h: u16) -> u16 {
-    viewport_h.saturating_sub(input_h + 1 + chooser_h + rail_h)
-}
-
-/// Rows the live transcript band may KEEP inside a viewport of `viewport_h`.
-///
-/// A grown composer is not counted: it is temporary, the operator is typing
-/// rather than reading, and rows flushed to fit it would leave a hole above
-/// the composer once it shrinks back. The chooser band IS counted, although
-/// it is temporary too: the operator must read the reply to choose, and a
-/// band that keeps rows it cannot show hides exactly that reply behind a
-/// "↑ 7 more" marker. Flushed rows land directly above the band, still on
-/// screen, so pushing the reply up costs a few blank rows after the pick
-/// and hiding it costs the reply. The rail stays for the session and takes
-/// its rows outright.
-fn band_inner_rows(app: &App, viewport_h: u16) -> u16 {
-    let chooser_h = chooser_band_height(app, viewport_h, INPUT_H_MIN);
-    band_capacity(viewport_h, INPUT_H_MIN, chooser_h, fleet_rail_height(app))
-}
-
-/// Index one past the last message that is settled AND therefore flushable:
-/// everything before the still-streaming turn.
-fn settle_end(app: &App) -> usize {
-    let total = app.messages.len();
-    let ceiling = if app.streaming {
-        total.saturating_sub(1)
-    } else {
-        total
-    };
-    let mut end = app.flushed_upto;
-    while end < ceiling && !app.messages[end].streaming {
-        end += 1;
-    }
-    end
-}
-
-fn prefix_hash(s: &str) -> u64 {
-    use std::hash::Hasher;
-    let mut h = std::hash::DefaultHasher::new();
-    h.write(s.as_bytes());
-    h.finish()
-}
-
-/// Bytes of `messages[flushed_upto].text` already committed to scrollback —
-/// or 0 when that bookkeeping no longer describes the message there.
-///
-/// It can stop describing it two ways: `finish_agent_turn` installs the
-/// authoritative reply over the streamed text, and `fail_turn` drops the
-/// streaming message outright. Both are caught by re-hashing the prefix, so a
-/// remainder is never spliced onto text that never had that prefix.
-fn effective_skip(app: &App) -> usize {
-    if app.flushed_bytes == 0 {
-        return 0;
-    }
-    let Some(m) = app.messages.get(app.flushed_upto) else {
-        return 0;
-    };
-    if m.role != Role::Agent || m.step.is_some() {
-        return 0;
-    }
-    match m.text.get(..app.flushed_bytes) {
-        Some(p) if prefix_hash(p) == app.flushed_hash => app.flushed_bytes,
-        _ => 0,
-    }
-}
-
-/// Byte offset one past the FIRST complete markdown block in `rest` (0 when
-/// there is none yet) — the next chunk of a streaming reply that can be
-/// committed to scrollback.
-///
-/// A block ends at a blank line outside a fenced code block. Committing whole
-/// blocks is what makes a chunk immutable: block-level markdown renders the
-/// same alone as it does in context, so the renderer can never want to rewrite
-/// a line that is already in scrollback and can no longer be redrawn.
-fn next_block_end(rest: &str) -> usize {
-    let mut in_fence = false;
-    let mut off = 0usize;
-    for line in rest.split_inclusive('\n') {
-        let trimmed = line.trim();
-        if trimmed.starts_with("```") || trimmed.starts_with("~~~") {
-            in_fence = !in_fence;
-        } else if !in_fence && trimmed.is_empty() {
-            return off + line.len();
-        }
-        off += line.len();
-    }
-    0
-}
-
-/// Lines the live band paints for one message, honoring a committed prefix on
-/// the band's head message (its committed part is already in scrollback).
-fn push_live(lines: &mut Vec<Line<'static>>, app: &App, m: &ChatMsg, skip: usize) {
-    push_live_inner(lines, app, m, skip, false)
-}
-
-/// Same, but rendered the way the message will look once it SETTLES — used for
-/// every flush decision.
-///
-/// A streaming body paints raw (markdown is only parsed at finish time), and
-/// raw is taller: markdown collapses the blank lines between list items and
-/// paragraphs. Measuring raw over-commits — the band would be full while
-/// streaming and then, the moment the turn settles and re-renders shorter,
-/// short by exactly the lines markdown folded away, with no way to pull them
-/// back out of scrollback. Measuring settled means the band is exactly full
-/// after the turn; while streaming it simply tail-follows, as it always has.
-/// Whether a gap row goes before this message.
-///
-/// A gap means "the speaker changed". A step card is a step *inside* the turn
-/// that precedes it, so a run of ten tool calls used to cost ten gap rows and
-/// read as ten unrelated events — twenty rows of scrollback for ten facts.
-/// Suppressing the gap there is what turns vertical rhythm into grouping
-/// instead of uniform repetition: every remaining gap now marks a real
-/// boundary.
-fn wants_gap_before(m: &super::app::ChatMsg) -> bool {
-    m.step.is_none()
-}
-
-/// The gap row itself — a rule when the skin asks for one AND the gap is a
-/// change of speaker between two spoken turns; otherwise a blank.
-///
-/// A System notice is the UI talking, not a speaker: three `/skin` switches
-/// used to draw three rules, and `/skills` output followed by its own notice
-/// read as four unrelated events. Notices group under a blank, and a rule is
-/// left meaning exactly one thing — the conversation moved to the other side.
-fn gap_row(
-    theme: &'static super::theme::Theme,
-    prev: Option<&super::app::ChatMsg>,
-    m: &super::app::ChatMsg,
-) -> Line<'static> {
-    let spoken = |r: Role| r != Role::System;
-    if theme.show_separator && prev.is_none_or(|p| spoken(p.role)) && spoken(m.role) {
-        Line::styled(
-            "─".repeat(SEPARATOR_WIDTH),
-            Style::default().fg(theme.separator),
-        )
-    } else {
-        Line::default()
-    }
-}
-
-/// The welcome while it is still the head of the live band: painted above
-/// the first message by `render_transcript`, and committed to scrollback
-/// with it by `flush_finished`. `None` once that head has been flushed (see
-/// [`App::welcome_header_live`]). Ends in a blank when there are messages
-/// under it, so the hint line and the first message do not touch.
-fn welcome_header(app: &App, eye_open: bool) -> Option<Vec<Line<'static>>> {
-    if !app.welcome_header_live() {
-        return None;
-    }
-    let mut lines = welcome_lines(
-        app.theme,
-        app.mascot_mode,
-        &app.agent,
-        app.cwd.as_deref(),
-        eye_open,
-    );
-    if !app.messages.is_empty() {
-        lines.push(Line::default());
-    }
-    Some(lines)
-}
-
-/// Lines for one message, including the gap that precedes it.
-///
-/// One function for all three callers — the viewport, the scrollback emit, and
-/// the row measurement that decides where to cut between them. They disagreed
-/// in two ways that both showed:
-///
-/// * the viewport drew no gaps at all, so the same transcript was dense on
-///   screen and spaced once it scrolled up — the reader's spatial map changed
-///   with no action from the reader;
-/// * the measurement summed message rows without the gap rows, so the band was
-///   taller than the number the flush decision was made from.
-///
-/// Attributing each gap to the message it precedes is what lets one function
-/// serve both: a measured block and a rendered block are the same rows.
-fn message_block(
-    app: &App,
-    idx: usize,
-    m: &super::app::ChatMsg,
-    skip: usize,
-    measured: bool,
-) -> Vec<Line<'static>> {
-    let mut lines: Vec<Line<'static>> = Vec::new();
-    // Never before a continuation: `skip > 0` resumes a message whose head is
-    // already committed above.
-    if idx > 0 && skip == 0 && wants_gap_before(m) {
-        lines.push(gap_row(app.theme, app.messages.get(idx - 1), m));
-    }
-    if measured {
-        push_live_measured(&mut lines, app, m, skip);
-    } else {
-        push_live(&mut lines, app, m, skip);
-    }
-    lines
-}
-
-fn push_live_measured(lines: &mut Vec<Line<'static>>, app: &App, m: &ChatMsg, skip: usize) {
-    push_live_inner(lines, app, m, skip, true)
-}
-
-fn push_live_inner(
-    lines: &mut Vec<Line<'static>>,
-    app: &App,
-    m: &ChatMsg,
-    skip: usize,
-    as_settled: bool,
-) {
-    let streaming_agent = m.streaming && m.role == Role::Agent && m.step.is_none();
-    if skip == 0 && !(as_settled && streaming_agent) {
-        push_message(
-            lines,
-            m,
-            app.spinner,
-            app.theme,
-            app.cards_expanded,
-            app.width,
-        );
-        return;
-    }
-    if skip == 0 {
-        // Measuring a streaming turn: header + reasoning, then the settled body.
-        push_agent_header(lines, m, app.spinner, app.theme);
-    }
-    // Continuation of a partially-committed agent turn: body only, no header.
-    lines.extend(agent_body_lines(
-        m.text.get(skip..).unwrap_or(""),
-        m.streaming && !as_settled,
-        app.spinner,
-        app.theme,
-        None,
-    ));
-}
-
-/// Wrapped (physical) row count of `lines` inside the band, measured exactly
-/// the way `render_transcript` paints them (`Paragraph::line_count`, not
-/// `lines.len()`), so flush decisions stay in lock-step with the band.
-/// `outer_width` is the full pane width, before the border block trims it.
-fn band_rows(
-    theme: &'static super::theme::Theme,
-    lines: Vec<Line<'static>>,
-    outer_width: u16,
-) -> u16 {
-    if lines.is_empty() {
-        return 0;
-    }
-    let block = Block::default().padding(Padding::horizontal(theme.inner_padding as u16));
-    let inner_width = block.inner(Rect::new(0, 0, outer_width.max(1), 1)).width;
-    Paragraph::new(Text::from(lines))
-        .wrap(Wrap { trim: false })
-        .line_count(inner_width.max(1)) as u16
-}
-
-/// Print `lines` into the terminal's scrollback above the inline viewport.
-fn emit<B: Backend>(
-    terminal: &mut ratatui::Terminal<B>,
-    lines: Vec<Line<'static>>,
-    pad: u16,
-    width: u16,
-) -> std::io::Result<()> {
-    // Height must be the WRAPPED (physical) row count, not the logical line
-    // count: `insert_before` renders into a buffer exactly `height` rows tall,
-    // and `Wrap` soft-wraps any line wider than the pane into extra rows. Using
-    // `lines.len()` clips every wrapped overflow row — a long message loses its
-    // tail into the void (never reaches scrollback, so it can't be scrolled
-    // back to). `Paragraph::line_count(width)` accounts for wrap + the padding
-    // block. (Enabled by the `unstable-rendered-line-info` ratatui feature.)
-    let text = Text::from(lines);
-    let block = || Block::default().padding(Padding::horizontal(pad));
-    let height = (Paragraph::new(text.clone())
-        .wrap(Wrap { trim: false })
-        .block(block())
-        .line_count(width) as u16)
-        .max(1);
-    terminal.insert_before(height, |buf| {
-        Paragraph::new(text)
-            .wrap(Wrap { trim: false })
-            .block(block())
-            .render(buf.area, buf);
-        blank_wide_char_continuations(buf);
-    })
-}
-
-/// Flush the OVERFLOW of the live band into the terminal's native scrollback
-/// via `Terminal::insert_before`: the oldest settled messages, then complete
-/// blocks of a still-streaming reply — and in both cases only as much as it
-/// takes for what remains to fit the band.
-///
-/// Why overflow and not "everything settled": the Inline viewport has a fixed
-/// height (see `viewport_h_for`), so flushing eagerly leaves the band blank and
-/// the composer sitting that many rows above the screen bottom. Keeping the
-/// band full of real content is what glues the composer to the bottom row with
-/// no gap, and it means the viewport is never resized — so none of ratatui's
-/// re-anchor paths (which can only leak blank rows into scrollback or float the
-/// viewport up) ever run.
-///
-/// A whole message is flushable once it can no longer change: not itself
-/// streaming, and not the trailing entry while a turn is in progress. A
-/// streaming reply spills block by block instead, so a long answer scrolls into
-/// native scrollback as it arrives rather than being trapped in the band.
-/// No-op in Fullscreen mode (the overlay reads `app.messages` directly) and
-/// while the band still fits.
-pub fn flush_finished<B: Backend>(
-    terminal: &mut ratatui::Terminal<B>,
-    app: &mut App,
-    viewport_h: u16,
-) -> std::io::Result<()> {
-    use super::app::RenderMode;
-    if app.render_mode != RenderMode::Inline {
-        return Ok(());
-    }
-    let theme = app.theme;
-    let pad = theme.inner_padding as u16;
-    let width = terminal.size()?.width.max(1);
-    let cap = u32::from(band_inner_rows(app, viewport_h));
-
-    let mut skip = effective_skip(app);
-    if app.flushed_bytes > 0 && skip == 0 {
-        // The committed prefix no longer belongs to the message sitting there;
-        // forget it and let that message flush whole. A visible duplicate of
-        // the partial text beats splicing a remainder onto the wrong body.
-        app.flushed_bytes = 0;
-    }
-
-    // ── 1. whole settled messages, oldest first ────────────────────────────
-    // Per-message row counts: wrapping is per line and the band draws no
-    // separators between messages, so the band total is their sum. Summing
-    // once keeps this O(n) instead of re-measuring the whole tail per
-    // candidate index (a resize resets `flushed_upto` to 0).
-    let start = app.flushed_upto.min(app.messages.len());
-    let mut rows: Vec<u16> = app.messages[start..]
-        .iter()
-        .enumerate()
-        .map(|(n, m)| {
-            let lines = message_block(app, start + n, m, if n == 0 { skip } else { 0 }, true);
-            band_rows(theme, lines, width)
-        })
-        .collect();
-    // The welcome is a header over the first notice, never a message of its
-    // own: it is measured with message 0 and leaves with it, so the band the
-    // flush decision was made from is the band that was painted.
-    let welcome = welcome_header(app, true);
-    if let Some(r) = rows.first_mut()
-        && let Some(w) = welcome.as_ref()
-    {
-        *r = r.saturating_add(band_rows(theme, w.clone(), width));
-    }
-    let mut total: u32 = rows.iter().map(|r| u32::from(*r)).sum();
-    let settled = settle_end(app);
-    let mut end = start;
-    // Flush until what remains fits, even when that leaves the band short. The
-    // alternative — keep the message whose departure would leave a blank slab,
-    // and let the band hide its surplus rows above the fold — traded a few
-    // empty rows for a reply the reader could not see ("↑ 7 more · PgUp" over
-    // the one answer they were asked to act on). A flushed reply sits directly
-    // above the band, on screen; a hidden one is gone until they page for it.
-    while end < settled && total > cap {
-        total -= u32::from(rows[end - start]);
-        end += 1;
-    }
-    if end > start {
-        let mut lines: Vec<Line<'static>> = welcome.unwrap_or_default();
-        for i in start..end {
-            let msg_skip = if i == start { skip } else { 0 };
-            lines.extend(message_block(app, i, &app.messages[i], msg_skip, false));
-        }
-        emit(terminal, lines, pad, width)?;
-        app.flushed_upto = end;
-        app.flushed_bytes = 0;
-        skip = 0;
-    }
-
-    // ── 2. still overflowing → spill complete blocks of the streaming turn ──
-    // One block at a time, re-measuring, so the band keeps painting a full
-    // screenful instead of emptying out mid-turn.
-    while total > cap {
-        let Some(m) = app.messages.get(app.flushed_upto) else {
-            break;
-        };
-        // ponytail: reasoning turns keep the whole-message flush. `thinking`
-        // renders above the body, so committing a body block first would strand
-        // any reasoning that arrives later after it in scrollback.
-        if !m.streaming || m.role != Role::Agent || m.step.is_some() || !m.thinking.is_empty() {
-            break;
-        }
-        let rest = m.text.get(skip..).unwrap_or("");
-        let block_end = next_block_end(rest);
-        if block_end == 0 {
-            break;
-        }
-        let chunk = &rest[..block_end];
-        let mut lines: Vec<Line<'static>> = Vec::new();
-        if skip == 0 {
-            if app.flushed_upto > 0 {
-                // This path only ever emits an agent turn's own body, which
-                // always opens a turn — no `wants_gap_before` test needed, but
-                // the row itself comes from the one builder.
-                lines.push(gap_row(theme, app.messages.get(app.flushed_upto - 1), m));
-            }
-            lines.push(Line::from(Span::styled(
-                "● agent".to_string(),
-                Style::default()
-                    .fg(theme.agent)
-                    .add_modifier(Modifier::BOLD),
-            )));
-        }
-        lines.extend(agent_body_lines(chunk, false, app.spinner, theme, None));
-        emit(terminal, lines, pad, width)?;
-        skip += block_end;
-        app.flushed_bytes = skip;
-        app.flushed_hash = prefix_hash(&app.messages[app.flushed_upto].text[..skip]);
-
-        let mut live = Vec::new();
-        push_live_measured(&mut live, app, &app.messages[app.flushed_upto], skip);
-        total = u32::from(band_rows(theme, live, width));
-    }
-    Ok(())
-}
-
-/// Work around a ratatui 0.29 bug: `Terminal::insert_before` flushes the whole
-/// buffer through `draw_lines`, which (unlike the normal diff-based flush) does
-/// NOT skip the trailing continuation cell of a wide (CJK) grapheme. That cell
-/// holds a space, so every wide char prints as "char " — spacing out CJK text
-/// in scrollback. Blank the continuation cell's symbol so the backend prints
-/// nothing there; the cursor has already advanced two columns for the wide
-/// char, so the next glyph lands correctly. (Live-viewport draws use the diff
-/// path and are unaffected.)
-fn blank_wide_char_continuations(buf: &mut ratatui::buffer::Buffer) {
-    use unicode_width::UnicodeWidthStr;
-    let area = buf.area;
-    for y in area.top()..area.bottom() {
-        let mut x = area.left();
-        while x + 1 < area.right() {
-            if buf[(x, y)].symbol().width() >= 2 {
-                buf[(x + 1, y)].set_symbol("");
-                x += 2;
-            } else {
-                x += 1;
-            }
-        }
-    }
-}
-
-/// Right-hand title for the transcript's top border, or `None` when nothing
-/// is hidden.
-///
-/// A band that silently drops the rows above it is indistinguishable from one
-/// that never had them — which is exactly how a reply behind the suggested-
-/// reply chooser reads as lost. `max_scroll` is the number of rows above the
-/// band; `scroll_back` is how many the operator has already walked up.
-fn scroll_marker(max_scroll: u16, scroll_back: u16) -> Option<String> {
-    if max_scroll == 0 {
-        return None;
-    }
-    Some(if scroll_back == 0 {
-        format!(" ↑ {max_scroll} more · PgUp ")
-    } else {
-        format!(" ↑ {} · PgDn to follow ", max_scroll - scroll_back)
-    })
-}
-
-/// Draws the *live* region only: everything at index `< app.flushed_upto` has
-/// already been flushed into the terminal's own scrollback (see
-/// `flush_finished`), so this is at most the one currently-streaming message
-/// — auto-following its tail as it grows, the same way `tail -f` does, rather
-/// than user-controlled paging (there's nothing left here to page through;
-/// full history lives in native scrollback, or the Ctrl+O overlay's own
-/// `scroll_back`-driven view of the complete `app.messages`).
-fn render_transcript(f: &mut Frame, app: &mut App, area: Rect) {
-    let theme = app.theme;
-    // No border at all.
-    //
-    // The name was already in the status-bar badge, so a titled frame spent a
-    // full-width rule restating it, and the bottom rule marked the same seam
-    // as the composer's own titled top border beneath it. The top rule went
-    // last: it only ever hosted the scroll marker, and on a screen already
-    // ruled between speakers and around the composer it read as one line too
-    // many. The marker now paints itself on the band's first row, right-
-    // aligned, when — and only when — rows are hidden above.
-    let block = Block::default().padding(Padding::horizontal(theme.inner_padding as u16));
-    let inner = block.inner(area);
-    let inner_width = inner.width;
-
-    let start = app.flushed_upto.min(app.messages.len());
-    // The head message may be partially committed to scrollback already (a
-    // streaming reply spills complete blocks); paint only what follows.
-    let skip = effective_skip(app);
-    let mut lines: Vec<Line> = Vec::new();
-    for (n, m) in app.messages[start..].iter().enumerate() {
-        lines.extend(message_block(
-            app,
-            start + n,
-            m,
-            if n == 0 { skip } else { 0 },
-            false,
-        ));
-    }
-
-    // No conversation yet → progressive-disclosure welcome (mascot + identity
-    // + one example + /help hint) above whatever notices slash commands have
-    // produced, instead of a bare prompt. The eye frame is a pure function of
-    // wall-clock time; the event loop schedules redraws on the blink deadline
-    // so an idle welcome animates without busy-looping.
-    if let Some(mut head) = welcome_header(app, app.blink.eye_open(Instant::now())) {
-        head.append(&mut lines);
-        lines = head;
-    }
-
-    // Same wrapped-row accounting as before (`line_count`, not `lines.len()`).
-    // `scroll_back` counts lines up from the bottom (0 = follow the tail);
-    // PageUp/PageDown and the chooser both rely on this so a long unflushed
-    // reply stays reachable even while an option list is open.
-    let output = Paragraph::new(Text::from(lines)).wrap(Wrap { trim: false });
-    let total = output.line_count(inner_width) as u16;
-    let visible = inner.height;
-    app.scroll_page = visible.max(1);
-    let max_scroll = total.saturating_sub(visible);
-    app.scroll_back = app.scroll_back.min(max_scroll);
-    let offset = max_scroll - app.scroll_back;
-
-    f.render_widget(output.block(block).scroll((offset, 0)), area);
-    if let Some(marker) = scroll_marker(max_scroll, app.scroll_back) {
-        let row = Rect { height: 1, ..inner };
-        f.render_widget(
-            Line::styled(marker, Style::default().fg(theme.border_title)).right_aligned(),
-            row,
-        );
-    }
-}
-
-/// Header line of an agent turn plus its reasoning block: an animated bullet
-/// while streaming, a solid one once done — so an in-progress turn reads as
-/// "live" at a glance vs. a finished one. Reasoning stays visible after the
-/// turn finishes (D5).
-fn push_agent_header(
-    lines: &mut Vec<Line<'static>>,
-    m: &ChatMsg,
-    spinner: usize,
-    theme: &'static super::theme::Theme,
-) {
-    let (bullet, header_style) = if m.streaming {
-        let spin = SPINNER[spinner % SPINNER.len()];
-        (format!("{spin} agent"), Style::default().fg(theme.agent))
-    } else {
-        (
-            "● agent".to_string(),
-            Style::default()
-                .fg(theme.agent)
-                .add_modifier(Modifier::BOLD),
-        )
-    };
-    lines.push(Line::from(Span::styled(bullet, header_style)));
-    if !m.thinking.is_empty() {
-        for l in m.thinking.lines() {
-            lines.push(Line::styled(
-                format!("{MSG_INDENT}{l}"),
-                Style::default()
-                    .fg(theme.thinking)
-                    .add_modifier(Modifier::ITALIC | Modifier::DIM),
-            ));
-        }
-    }
-}
-
-/// Body lines of an agent turn — no header, no reasoning: raw text plus a
-/// trailing spinner while streaming, markdown-rendered once settled.
-///
-/// `cached` is the markdown rendered once at finish time for the WHOLE message
-/// (`ChatMsg::rendered`); pass `None` when rendering a slice of it, so the
-/// slice gets its own render instead of the whole reply's.
-fn agent_body_lines(
-    text: &str,
-    streaming: bool,
-    spinner: usize,
-    theme: &'static super::theme::Theme,
-    cached: Option<&Vec<Line<'static>>>,
-) -> Vec<Line<'static>> {
-    if streaming {
-        let mut body: Vec<Line<'static>> = text
-            .lines()
-            .map(|l| Line::raw(format!("{MSG_INDENT}{l}")))
-            .collect();
-        // Trailing spinner so the user sees liveness.
-        let spin = SPINNER[spinner % SPINNER.len()];
-        match body.last_mut() {
-            Some(last) => last.spans.push(Span::styled(
-                format!(" {spin}"),
-                Style::default().fg(theme.agent),
-            )),
-            None => body.push(Line::styled(
-                spin.to_string(),
-                Style::default().fg(theme.agent),
-            )),
-        }
-        body
-    } else if let Some(cached) = cached {
-        // Finished reply: reuse the markdown rendered once at finish time.
-        cached.iter().cloned().map(indent_line).collect()
-    } else {
-        markdown::render(text)
-            .lines
-            .into_iter()
-            .map(indent_line)
-            .collect()
-    }
-}
-
-fn push_message(
-    lines: &mut Vec<Line<'static>>,
-    m: &ChatMsg,
-    spinner: usize,
-    theme: &'static super::theme::Theme,
-    cards_expanded: bool,
-    width: u16,
-) {
-    // Step cards replace role-based rendering entirely for that message.
-    if let Some(card) = &m.step {
-        lines.extend(super::render_card::card_lines(
-            card,
-            theme,
-            cards_expanded,
-            width,
-        ));
-        return;
-    }
-    match m.role {
-        Role::User => {
-            lines.push(Line::from(Span::styled(
-                "you ›",
-                Style::default().fg(theme.user).add_modifier(Modifier::BOLD),
-            )));
-            for l in m.text.lines() {
-                lines.push(Line::styled(
-                    format!("{MSG_INDENT}{l}"),
-                    Style::default().fg(theme.user_text),
-                ));
-            }
-        }
-        Role::System => {
-            // Severity paints the whole note and picks a lead glyph, so a
-            // warning reads amber and a success reads green at a glance.
-            let (color, glyph) = match m.severity {
-                Severity::Info => (theme.system, "·"),
-                Severity::Warn => (theme.warn, "▲"),
-                Severity::Error => (theme.error, "✖"),
-                Severity::Success => (theme.success, "✔"),
-            };
-            let bold = !matches!(m.severity, Severity::Info);
-            for (i, l) in m.text.lines().enumerate() {
-                let prefix = if i == 0 {
-                    format!("{glyph} ")
-                } else {
-                    "  ".to_string()
-                };
-                let mut style = Style::default().fg(color);
-                if bold {
-                    style = style.add_modifier(Modifier::BOLD);
-                }
-                lines.push(Line::styled(format!("{prefix}{l}"), style));
-            }
-        }
-        Role::Shell => {
-            // `$ cmd` highlighted, output dim — visually a local terminal block.
-            let mut it = m.text.lines();
-            if let Some(first) = it.next() {
-                lines.push(Line::styled(
-                    first.to_string(),
-                    Style::default()
-                        .fg(theme.shell)
-                        .add_modifier(Modifier::BOLD),
-                ));
-            }
-            for l in it {
-                lines.push(Line::styled(
-                    l.to_string(),
-                    Style::default().fg(theme.system),
-                ));
-            }
-        }
-        Role::Agent => {
-            // Animated bullet while streaming, solid bullet once done — so an
-            // in-progress turn reads as "live" at a glance vs. a finished one.
-            push_agent_header(lines, m, spinner, theme);
-            lines.extend(agent_body_lines(
-                &m.text,
-                m.streaming,
-                spinner,
-                theme,
-                m.rendered.as_ref(),
-            ));
-            // Drawn here, not baked into `rendered`: this is the only place
-            // that knows the pane width, and it runs every frame, so the card
-            // reflows on resize for free.
-            if let Some(body) = &m.settlement {
-                let inner = width.saturating_sub(u16::from(theme.inner_padding) * 2);
-                lines.extend(super::settlement::card_lines(body, theme, inner));
-            }
-        }
-    }
-}
-
-/// Separator between status-bar segments.
-const FOOTER_SEP: &str = " · ";
-
-/// Below this many columns the status bar drops the steering hint and keeps
-/// the numbers.
-const STATUS_FULL_MIN_WIDTH: u16 = 100;
-
-/// Longest joined tool-name list the `AUTO:` badge will spell out before it
-/// falls back to a bare count. Sized so the badge cannot crowd out the rest of
-/// the status bar on a narrow terminal — two typical tool names fit.
-const AUTO_NAMES_MAX: usize = 24;
-
-fn render_status(f: &mut Frame, app: &App, area: Rect) {
-    let theme = app.theme;
-    let (msg, color) = if let Some(req) = &app.hitl {
-        // Surface the auto-deny clock: the gate expires approvals after
-        // DEFAULT_TIMEOUT (300s). Reuse `created_at` rather than tracking new
-        // state; the status bar redraws on each blink deadline so it ticks.
-        // The tool name goes here so the operator always knows WHAT they are
-        // being asked to approve at any width; the decision keys live in the
-        // framed modal / inline row, which is where the user is looking.
-        let remaining = crate::hitl::gate::DEFAULT_TIMEOUT
-            .as_secs()
-            .saturating_sub(req.created_at.elapsed().as_secs());
-        (
-            format!("⏳ approve {} · auto-deny in {remaining}s", req.tool_name),
-            Color::Yellow,
-        )
-    } else if app.streaming {
-        let spin = SPINNER[app.spinner % SPINNER.len()];
-        // The steering hint is the first thing to drop when the row is tight:
-        // it is advice, and everything to its right is state.
-        let msg = if area.width >= STATUS_FULL_MIN_WIDTH {
-            format!("{spin} generating… · type to steer · Ctrl+C to cancel")
-        } else {
-            format!("{spin} generating…")
-        };
-        (msg, theme.agent)
-    } else {
-        let ctx = if app.context_task_id.is_some() {
-            " · context kept"
-        } else {
-            ""
-        };
-        (format!("ready{ctx}"), theme.system)
-    };
-    let mut spans = vec![
-        Span::styled(
-            format!(" {} ", app.agent),
-            Style::default().fg(theme.badge_fg).bg(theme.badge_bg),
-        ),
-        Span::raw("  "),
-    ];
-    // Auto-approval visibility (#8 / proposal 2). All three auto-approval
-    // paths now show a badge, not just the global `auto_approve`:
-    //   - `auto_approve`               → ` AUTO `   (every tool, session)
-    //   - `session_tool_allow` (N>0)   → ` AUTO:N ` (N tools muted via [a])
-    //   - `auto_reads`                 → ` READS `  (read_file auto-approved)
-    // Pure display; no behaviour change. Fixes the "AUTO badge vanished in a
-    // new session" illusion where `[a]`-muted tools left no visible trace.
-    if app.auto_approve {
-        spans.push(Span::styled(
-            " AUTO ",
-            Style::default()
-                .fg(Color::Black)
-                .bg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ));
-        spans.push(Span::raw("  "));
-    } else if !app.session_tool_allow.is_empty() {
-        // Name the muted tools when they fit. `AUTO:2` said something was
-        // muted but never what, and nothing else would tell you either —
-        // so the operator could not check whether a grant they did not mean
-        // to make was still in force. `/auto off` revokes them.
-        let mut names: Vec<&str> = app.session_tool_allow.iter().map(String::as_str).collect();
-        names.sort_unstable();
-        let joined = names.join(",");
-        let label = if joined.chars().count() <= AUTO_NAMES_MAX {
-            format!(" AUTO:{joined} ")
-        } else {
-            format!(" AUTO:{} ", names.len())
-        };
-        spans.push(Span::styled(
-            label,
-            Style::default()
-                .fg(Color::Black)
-                .bg(Color::Yellow)
-                .add_modifier(Modifier::BOLD),
-        ));
-        spans.push(Span::raw("  "));
-    }
-    if app.auto_reads {
-        spans.push(Span::styled(
-            " READS ",
-            Style::default()
-                .fg(Color::Black)
-                .bg(Color::Cyan)
-                .add_modifier(Modifier::BOLD),
-        ));
-        spans.push(Span::raw("  "));
-    }
-    if let Some(meta) = &app.channel {
-        // Id only. The chip used to append `meta.state`, a persisted channel
-        // lifecycle word refreshed at just two points, next to the live turn
-        // state a few columns to its right — so the bar routinely read
-        // `⏵ 019ff831:working   ready`. Two sources for one fact; the live one
-        // wins and the stale one is gone (#940).
-        let short: String = meta.id.chars().take(8).collect();
-        spans.push(Span::styled(
-            format!(" ⏵ {short} "),
-            Style::default().fg(theme.agent),
-        ));
-        spans.push(Span::raw("  "));
-    }
-    spans.push(Span::styled(msg, Style::default().fg(color)));
-
-    // Glass Box observability: tokens · cost · ctx · timer.
-    //
-    // Budgeted against the room actually left on the row. This used to be
-    // assembled at full length and handed to the terminal, which clipped the
-    // overflow — and what fell off the right edge was the context bar, the one
-    // figure here that changes what you do next.
-    let timer = app.turn_started.map(|t0| {
-        let secs = t0.elapsed().as_secs();
-        format!("{}m{:02}s · esc=stop", secs / 60, secs % 60)
-    });
-    let reserved: usize = spans
-        .iter()
-        .map(|s| s.content.chars().count())
-        .sum::<usize>()
-        + timer
-            .as_deref()
-            .map_or(0, |t| t.chars().count() + FOOTER_SEP.chars().count());
-    let obs = footer_segments(
-        app.turn_in,
-        app.turn_out,
-        app.session_in,
-        app.session_out,
-        app.ctx_tokens,
-        &app.pricing,
-        app.budget_usd,
-        usize::from(area.width).saturating_sub(reserved + FOOTER_SEP.chars().count()),
-    );
-    if !obs.is_empty() {
-        spans.push(Span::raw(FOOTER_SEP));
-        spans.push(Span::styled(obs, Style::default().fg(theme.system)));
-    }
-    if let Some(t) = timer {
-        spans.push(Span::styled(
-            format!("{FOOTER_SEP}{t}"),
-            Style::default().fg(theme.system),
-        ));
-    }
-
-    let right_hint: Option<(String, Color)> = if app.scroll_back > 0 {
-        Some((
-            format!("↑ {} lines · ⬇ to bottom", app.scroll_back),
-            theme.system,
-        ))
-    } else if app.esc_hint {
-        let hint = if app.streaming {
-            "ESC again to cancel"
-        } else {
-            "ESC again to clear"
-        };
-        Some((hint.to_string(), theme.system))
-    } else if app.ctrl_c_hint {
-        Some(("Ctrl+C again to quit".to_string(), theme.system))
-    } else {
-        None
-    };
-
-    if let Some((hint_text, hint_color)) = right_hint {
-        let hint_display = format!(" {} ", hint_text);
-        let hint_width = hint_display.chars().count() as u16;
-        let left_width: u16 = spans.iter().map(|s| s.content.chars().count() as u16).sum();
-        let bar_width = area.width;
-        // Only right-align if there's room; otherwise skip the hint.
-        if bar_width > left_width + hint_width {
-            let pad = bar_width - left_width - hint_width;
-            spans.push(Span::raw(" ".repeat(pad as usize)));
-            spans.push(Span::styled(
-                hint_display,
-                Style::default().fg(hint_color).add_modifier(Modifier::DIM),
-            ));
-        }
-    }
-
-    f.render_widget(Paragraph::new(Line::from(spans)), area);
-}
-
-/// Hard-wrap `s` into rows at most `w` display columns wide.
-///
-/// Character-based rather than word-based: the payload is pretty-printed JSON
-/// and shell, where breaking mid-token is honest and dropping the tail is not.
-/// Widths come from `unicode_width` so a CJK argument does not overflow the
-/// border.
-fn wrap_row(s: &str, w: usize) -> Vec<String> {
-    use unicode_width::UnicodeWidthChar;
-    if w == 0 || s.is_empty() {
-        return vec![s.to_string()];
-    }
-    let mut rows = Vec::new();
-    let mut cur = String::new();
-    let mut used = 0usize;
-    for ch in s.chars() {
-        let cw = ch.width().unwrap_or(0);
-        if used + cw > w && !cur.is_empty() {
-            rows.push(std::mem::take(&mut cur));
-            used = 0;
-        }
-        cur.push(ch);
-        used += cw;
-    }
-    rows.push(cur);
-    rows
-}
-
-/// Draw the approval modal. Returns the scroll offset it actually used —
-/// `scroll` clamped to the content, so the caller's stored offset cannot run
-/// away past the end of a short input — and how many body rows it had room to
-/// display, which is what the key handler pages by (see `hitl_scroll_step`).
-fn render_hitl(
-    f: &mut Frame,
-    hitl: &super::stream::HitlRequest,
-    grant_confirm: Option<char>,
-    composer_empty: bool,
-    scroll: u16,
-) -> (u16, u16) {
-    let area = centered_rect(HITL_PCT_X, HITL_PCT_Y, f.area());
-    let input = serde_json::to_string_pretty(&hitl.tool_input).unwrap_or_default();
-    // Header rows stay pinned: scrolling the body must never carry the tool
-    // name off-screen, since "which tool" is half of what is being approved.
-    let head = vec![
-        Line::from(Span::styled(
-            hitl.prompt.clone(),
-            Style::default().add_modifier(Modifier::BOLD),
-        )),
-        Line::default(),
-        Line::from(vec![
-            Span::styled("tool: ", Style::default().fg(Color::DarkGray)),
-            Span::styled(hitl.tool_name.clone(), Style::default().fg(Color::Yellow)),
-        ]),
-    ];
-    let row_w = area.width.saturating_sub(2) as usize;
-    // Wrap every line and keep every line (#939). This modal exists so a human
-    // reads the command before it runs; a destructive suffix past a horizontal
-    // cut, or past a `.take(12)`, is exactly what must not be silently dropped.
-    let mut body: Vec<Line> = Vec::new();
-    for l in input.lines() {
-        for row in wrap_row(l, row_w) {
-            body.push(Line::styled(row, Style::default().fg(Color::DarkGray)));
-        }
-    }
-    // When a session-wide grant is armed, the modal shows ONLY the confirm
-    // instruction: the operator is answering "do you really mean the whole
-    // session?", and re-printing the full key row there invites a reflex press.
-    let keys = if let Some(c) = grant_confirm {
-        let what = if c == 'a' {
-            format!("`{}` for this session", hitl.tool_name)
-        } else {
-            "ALL tools for this session".to_string()
-        };
-        Line::from(vec![
-            Span::styled(
-                format!("press [{c}] again"),
-                Style::default()
-                    .fg(Color::Magenta)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::raw(format!(" to allow {what} — any other key cancels")),
-        ])
-    } else {
-        Line::from(vec![
-            Span::styled(
-                "[y]",
-                Style::default()
-                    .fg(Color::Green)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::raw(" approve    "),
-            Span::styled(
-                "[a]",
-                Style::default()
-                    .fg(Color::Yellow)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::raw(" always allow this tool (session)    "),
-            Span::styled(
-                "[A]",
-                Style::default()
-                    .fg(Color::Magenta)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::raw(" allow all tools (session)    "),
-            Span::styled(
-                "[n]",
-                Style::default().fg(Color::Red).add_modifier(Modifier::BOLD),
-            ),
-            Span::raw(" deny / Esc"),
-        ])
-    };
-
-    let block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(Style::default().fg(Color::Yellow))
-        .title(" approve tool call ");
-    let inner = block.inner(area);
-    f.render_widget(Clear, area);
-    f.render_widget(block, area);
-
-    // The key row is the only part of this modal that is never optional, so it
-    // gets its own chunk. Previously it was the last entry in one clipped
-    // Paragraph: a wrapped JSON input pushed it out of the box and left the
-    // operator staring at a blocking gate with no visible way to answer it.
-    //
-    // While the composer holds text the `composer_empty` guard (#893) makes
-    // y/a/A/n type instead of decide. Say so, and dim the row: advertising a
-    // live key that is inert is how an operator ends up hitting the 5-minute
-    // auto-deny wondering why nothing responds (#939).
-    let keys_inert = !composer_empty && grant_confirm.is_none();
-    let keys_text = if keys_inert {
-        let dimmed = Line::from(
-            keys.spans
-                .iter()
-                .map(|s| {
-                    Span::styled(
-                        s.content.clone(),
-                        s.style.add_modifier(Modifier::DIM).fg(Color::DarkGray),
-                    )
-                })
-                .collect::<Vec<_>>(),
-        );
-        Text::from(vec![
-            dimmed,
-            Line::styled(
-                "these keys type while the composer has text — Ctrl+U clears it",
-                Style::default().fg(Color::Yellow),
-            ),
-        ])
-    } else {
-        Text::from(keys)
-    };
-    let keys_h = Paragraph::new(keys_text.clone())
-        .wrap(Wrap { trim: false })
-        .line_count(inner.width.max(1)) as u16;
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Min(0), Constraint::Length(keys_h.max(1))])
-        .split(inner);
-
-    // Body = pinned header + a scrolled window over the wrapped input. When it
-    // does not all fit, the notice reports what is HIDDEN — the old message
-    // printed the number of rows kept, so widening the pane made the "hidden"
-    // count go up (#939).
-    let body_h = chunks[0].height as usize;
-    let room = body_h.saturating_sub(head.len());
-    let mut lines = head;
-    let mut shown = room;
-    let used_scroll = if body.len() > room && room > 1 {
-        let visible = room - 1;
-        shown = visible;
-        let above = (scroll as usize).min(body.len() - visible);
-        let below = body.len() - visible - above;
-        lines.extend(body.into_iter().skip(above).take(visible));
-        let note = if above == 0 {
-            format!("… {below} more lines — PgDn to scroll")
-        } else {
-            format!("… {above} above · {below} below — PgUp/PgDn")
-        };
-        lines.push(Line::styled(note, Style::default().fg(Color::DarkGray)));
-        above as u16
-    } else {
-        lines.extend(body);
-        0
-    };
-    f.render_widget(Paragraph::new(Text::from(lines)), chunks[0]);
-    f.render_widget(
-        Paragraph::new(keys_text).wrap(Wrap { trim: false }),
-        chunks[1],
-    );
-    (used_scroll, shown as u16)
-}
-
-/// Format a token count with thousands separator (e.g. 1240 → "1,240").
-fn fmt_tok(n: u64) -> String {
-    if n >= 1_000_000 {
-        format!(
-            "{},{:03},{:03}",
-            n / 1_000_000,
-            (n / 1_000) % 1_000,
-            n % 1_000
-        )
-    } else if n >= 1_000 {
-        format!("{},{:03}", n / 1_000, n % 1_000)
-    } else {
-        n.to_string()
-    }
-}
-
-/// Pure footer formatter: `"{turn}/{sess} tok · {cost} · ctx <bar> N%"`,
-/// trimmed to `budget` columns.
-///
-/// Cost is the SESSION estimate, not the turn's: the turn's usage is zero for
-/// the whole time a reply is streaming, so the bar read `$0.000 est` next to a
-/// six-figure session token count — which parses as "this was free". When a cap
-/// is set the same figure renders as `spent / cap` instead of printing two
-/// different costs side by side. `—` when the model is unpriced; the ctx part
-/// is omitted when no window is known.
-///
-/// Segments drop right-to-left as `budget` shrinks — token pair first, then
-/// cost — because the context bar is the only number here that changes what
-/// you do next.
-#[allow(clippy::too_many_arguments)]
-fn footer_segments(
-    turn_in: u64,
-    turn_out: u64,
-    sess_in: u64,
-    sess_out: u64,
-    ctx_tokens: u64,
-    pricing: &super::footer::Pricing,
-    budget_usd: Option<f64>,
-    budget: usize,
-) -> String {
-    use super::footer::{CTX_BAR_WIDTH, UsageCounts, context_pct, ctx_bar, turn_cost};
-
-    let toks = format!(
-        "{}/{} tok",
-        fmt_tok(turn_in + turn_out),
-        fmt_tok(sess_in + sess_out)
-    );
-
-    let spent = turn_cost(
-        pricing,
-        &UsageCounts {
-            input: sess_in,
-            output: sess_out,
-        },
-    );
-    let cost = match (spent, budget_usd) {
-        (Some(c), Some(cap)) => format!("${c:.2} / ${cap:.2}"),
-        (Some(c), None) => format!("${c:.3} est"),
-        (None, Some(cap)) => format!("/ ${cap:.2}"),
-        // No per-token price for this model (local runtimes, and any registry
-        // entry added without `--input-cost`). An em dash in the same slot as
-        // "$0.125 est" read as "this turn was free" rather than "we cannot
-        // price it" — same field, two meanings, no way to tell them apart
-        // (#940).
-        (None, None) => "no price".to_string(),
-    };
-
-    let ctx = match pricing.window {
-        Some(w) if w > 0 => {
-            let pct = context_pct(ctx_tokens, w);
-            format!("ctx {} {}%", ctx_bar(pct, CTX_BAR_WIDTH), pct)
-        }
-        _ => String::new(),
-    };
-
-    for parts in [
-        [toks.as_str(), cost.as_str(), ctx.as_str()].as_slice(),
-        [cost.as_str(), ctx.as_str()].as_slice(),
-        [ctx.as_str()].as_slice(),
-    ] {
-        let s = parts
-            .iter()
-            .filter(|p| !p.is_empty())
-            .copied()
-            .collect::<Vec<_>>()
-            .join(FOOTER_SEP);
-        if s.chars().count() <= budget {
-            return s;
-        }
-    }
-    // Nothing fits: print nothing rather than a mangled half-number.
-    String::new()
-}
-
-fn centered_rect(pct_x: u16, pct_y: u16, area: Rect) -> Rect {
-    let v = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Percentage((100 - pct_y) / 2),
-            Constraint::Percentage(pct_y),
-            Constraint::Percentage((100 - pct_y) / 2),
-        ])
-        .split(area);
-    Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([
-            Constraint::Percentage((100 - pct_x) / 2),
-            Constraint::Percentage(pct_x),
-            Constraint::Percentage((100 - pct_x) / 2),
-        ])
-        .split(v[1])[1]
-}
-
-#[cfg(test)]
-mod block_boundary_tests {
-    use super::next_block_end;
-
-    #[test]
-    fn commits_one_complete_paragraph_at_a_time() {
-        let s = "first para\n\nsecond para\n\nthird";
-        let n = next_block_end(s);
-        assert_eq!(&s[..n], "first para\n\n");
-        // The next call, on the remainder, takes exactly the next block.
-        let rest = &s[n..];
-        let m = next_block_end(rest);
-        assert_eq!(&rest[..m], "second para\n\n");
-        // A trailing block with no blank line after it is never committable —
-        // it may still grow.
-        assert_eq!(next_block_end(&rest[m..]), 0);
-    }
-
-    #[test]
-    fn a_blank_line_inside_a_fence_is_not_a_boundary() {
-        // Committing here would strand an unterminated ``` in scrollback and
-        // re-render the fence body once it closes.
-        let s = "```rust\nlet a = 1;\n\nlet b = 2;\n```\n\ntail";
-        let n = next_block_end(s);
-        assert_eq!(&s[..n], "```rust\nlet a = 1;\n\nlet b = 2;\n```\n\n");
-        assert_eq!(next_block_end("```\nunclosed\n\nstill inside\n"), 0);
-    }
-
-    #[test]
-    fn nothing_to_commit_yet() {
-        assert_eq!(next_block_end(""), 0);
-        assert_eq!(next_block_end("one line, still streaming"), 0);
-    }
-}
-
-#[cfg(test)]
-mod wide_char_tests {
-    use super::blank_wide_char_continuations;
-    use ratatui::buffer::Buffer;
-    use ratatui::layout::Rect;
-    use ratatui::widgets::{Paragraph, Widget, Wrap};
-
-    #[test]
-    fn cjk_continuation_cells_are_blanked() {
-        let area = Rect::new(0, 0, 20, 1);
-        let mut buf = Buffer::empty(area);
-        Paragraph::new("有既有workflow")
-            .wrap(Wrap { trim: false })
-            .render(area, &mut buf);
-        // ratatui fills wide-char continuation cells with a space.
-        assert_eq!(buf[(1, 0)].symbol(), " ");
-        blank_wide_char_continuations(&mut buf);
-        // After the fix: each CJK glyph is followed by an empty (skipped) cell,
-        // so the backend prints "有既有" with no interleaved spaces.
-        assert_eq!(buf[(0, 0)].symbol(), "有");
-        assert_eq!(buf[(1, 0)].symbol(), "");
-        assert_eq!(buf[(2, 0)].symbol(), "既");
-        assert_eq!(buf[(3, 0)].symbol(), "");
-        assert_eq!(buf[(4, 0)].symbol(), "有");
-        assert_eq!(buf[(5, 0)].symbol(), "");
-        // ASCII run is untouched.
-        assert_eq!(buf[(6, 0)].symbol(), "w");
-        assert_eq!(buf[(7, 0)].symbol(), "o");
-    }
-}
-
-#[cfg(test)]
-mod footer_fmt_tests {
-    use super::footer_segments;
-    use crate::cmd::agent::cli::footer::Pricing;
-
-    /// Enough room that these assertions test the formatting, not the trimming.
-    const WIDE: usize = 200;
-
-    #[test]
-    fn shows_tokens_and_names_the_reason_when_unpriced() {
-        let s = footer_segments(1240, 0, 1240, 0, 0, &Pricing::default(), None, WIDE);
-        assert!(s.contains("1,240 tok") || s.contains("1240 tok"));
-        // Was an em dash, which sat in the same slot as "$0.125 est" and so
-        // read as a cost of zero rather than an absent price (#940).
-        assert!(s.contains("no price"), "got: {s}");
-    }
-
-    #[test]
-    fn shows_cost_and_ctx_when_priced() {
-        let p = Pricing {
-            in_per_1k: Some(0.003),
-            out_per_1k: Some(0.015),
-            window: Some(100_000),
-        };
-        let s = footer_segments(1000, 1000, 1000, 1000, 32_000, &p, None, WIDE);
-        assert!(s.contains("$0.018"));
-        assert!(s.contains("32%"));
-        assert!(!s.contains(" / $")); // no budget suffix when budget is None
-    }
-
-    #[test]
-    fn shows_budget_suffix_when_cap_set() {
-        let p = Pricing {
-            in_per_1k: Some(3.0),
-            out_per_1k: Some(15.0),
-            window: None,
-        };
-        // session 1000/1000 → $18.00 spent, cap $20.00
-        let s = footer_segments(1000, 1000, 1000, 1000, 0, &p, Some(20.0), WIDE);
-        assert!(s.contains("$18.00 / $20.00"), "got: {s}");
-        // unpriced model → spent omitted, cap still shown
-        let s2 = footer_segments(
-            1000,
-            1000,
-            1000,
-            1000,
-            0,
-            &Pricing::default(),
-            Some(20.0),
-            WIDE,
-        );
-        assert!(s2.contains("/ $20.00"), "got: {s2}");
-        assert!(!s2.contains("$18.00"));
-    }
-
-    #[test]
-    fn cost_is_the_session_not_the_turn() {
-        let p = Pricing {
-            in_per_1k: Some(3.0),
-            out_per_1k: Some(15.0),
-            window: None,
-        };
-        // Mid-stream: the turn's usage hasn't been reported yet. The bar used
-        // to price the turn, so it read "$0.000 est" beside a session total.
-        let s = footer_segments(0, 0, 1000, 1000, 0, &p, None, WIDE);
-        assert!(s.contains("$18.000 est"), "got: {s}");
-    }
-
-    #[test]
-    fn narrow_bar_keeps_the_context_gauge_and_drops_the_rest() {
-        let p = Pricing {
-            in_per_1k: Some(0.003),
-            out_per_1k: Some(0.015),
-            window: Some(100_000),
-        };
-        let full = footer_segments(1000, 1000, 1000, 1000, 91_000, &p, None, WIDE);
-        let tight = footer_segments(1000, 1000, 1000, 1000, 91_000, &p, None, 20);
-        assert!(full.len() > tight.len());
-        assert!(tight.contains("91%"), "got: {tight}");
-        assert!(!tight.contains("tok"), "got: {tight}");
-        assert!(tight.chars().count() <= 20, "got: {tight}");
-    }
-}
-
-#[cfg(test)]
-mod fleet_rail_layout_tests {
-    use super::*;
-    use crate::cmd::agent::cli::fleet_rail::{MemberRow, MemberState, RailView};
-
-    fn view(blocked: usize) -> RailView {
-        RailView {
-            jobs_line: "fleet · dev   job 0/1".into(),
-            members: (0..blocked)
-                .map(|i| MemberRow {
-                    agent: format!("m{i}"),
-                    state: MemberState::Blocked {
-                        summary: "approve".into(),
-                        hitl_id: format!("h{i}"),
-                    },
-                })
-                .collect(),
-            notice: None,
-        }
-    }
-
-    #[test]
-    fn rail_is_one_row_until_someone_is_blocked() {
-        assert_eq!(rail_height_for(&view(0)), 1);
-        assert_eq!(rail_height_for(&view(1)), 2);
-        assert_eq!(rail_height_for(&view(3)), 4);
-    }
-
-    #[test]
-    fn the_expanded_rail_is_capped() {
-        use crate::cmd::agent::cli::fleet_rail::MAX_EXPANDED_ROWS;
-        assert_eq!(
-            rail_height_for(&view(50)),
-            // +1 for the member rows (capped), +1 for the "… N more" notice
-            // that only appears once the list is actually truncated.
-            1 + MAX_EXPANDED_ROWS as u16 + 1,
-            "an unbounded rail would eat the transcript, and a truncated one \
-             must still show its own truncation notice"
-        );
-    }
-
-    #[test]
-    fn the_live_band_gives_back_exactly_what_the_rail_takes() {
-        // The guard for the one dangerous coupling: band_inner_rows decides
-        // when transcript content is flushed to scrollback, so it must account
-        // for every row the rail paints or the flush drifts from the picture.
-        // Tested on the pure arithmetic so no App and no test-only seam in
-        // production code are needed.
-        let viewport_h = 20u16;
-        let input_h = 3u16;
-        let without = band_capacity(viewport_h, input_h, 0, 0);
-        let with_rail = band_capacity(viewport_h, input_h, 0, rail_height_for(&view(3)));
-        assert_eq!(without - with_rail, rail_height_for(&view(3)));
-    }
-
-    #[test]
-    fn the_rail_height_matches_what_render_actually_paints() {
-        // rail_height_for is a number computed independently of rail_lines;
-        // nothing ties them together except this test. Without it, a line
-        // added to (or removed from) rail_lines silently desyncs the height
-        // from the paint — exactly how the truncation notice went missing.
-        use crate::cmd::agent::cli::fleet_rail::MAX_EXPANDED_ROWS;
-        let theme = crate::cmd::agent::cli::theme::resolve_skin("dark");
-        for blocked in [0, 1, MAX_EXPANDED_ROWS, MAX_EXPANDED_ROWS + 3] {
-            let v = view(blocked);
-            assert_eq!(
-                rail_lines(&v, theme).len() as u16,
-                rail_height_for(&v),
-                "mismatch at {blocked} blocked members"
-            );
-        }
-    }
-}
-
-#[cfg(test)]
-mod settlement_paint_tests {
-    use super::super::app::{ChatMsg, Role};
-    use super::super::theme::DARK;
-    use super::push_message;
-
-    #[test]
-    fn a_carried_settlement_is_painted_after_the_body() {
-        let mut m = ChatMsg::for_test(Role::Agent, "did it");
-        m.settlement = Some("  ✔ bash · cargo test".into());
-        let mut lines = Vec::new();
-        push_message(&mut lines, &m, 0, &DARK, false, 60);
-        let text: Vec<String> = lines
-            .iter()
-            .map(|l| l.spans.iter().map(|s| s.content.as_ref()).collect())
-            .collect();
-        assert!(
-            text.iter().any(|l| l.contains("SETTLEMENT")),
-            "card missing: {text:?}"
-        );
-        assert!(
-            text.iter().any(|l| l.contains("cargo test")),
-            "card body missing: {text:?}"
-        );
-    }
-
-    #[test]
-    fn a_message_without_one_paints_nothing_extra() {
-        let m = ChatMsg::for_test(Role::Agent, "did it");
-        let mut lines = Vec::new();
-        push_message(&mut lines, &m, 0, &DARK, false, 60);
-        let text: Vec<String> = lines
-            .iter()
-            .map(|l| l.spans.iter().map(|s| s.content.as_ref()).collect())
-            .collect();
-        assert!(!text.iter().any(|l| l.contains("SETTLEMENT")), "{text:?}");
-    }
-}
-
-#[cfg(test)]
-mod hitl_modal_tests {
-    use super::super::stream::HitlRequest;
-    use super::render_hitl;
-    use ratatui::Terminal;
-    use ratatui::backend::TestBackend;
-
-    /// A tool input big enough that, wrapped, it fills the modal on its own.
-    fn fat_request() -> HitlRequest {
-        HitlRequest {
-            hitl_id: "h1".into(),
-            step_id: None,
-            tool_name: "bash".into(),
-            tool_input: serde_json::json!({
-                "command": "x".repeat(400),
-                "cwd": "/Volumes/Firecuda4tb/Projects/mur",
-            }),
-            prompt: "Run `bash`?".into(),
-            created_at: std::time::Instant::now(),
-        }
-    }
-
-    #[test]
-    fn the_key_row_survives_an_oversized_input() {
-        let mut term = Terminal::new(TestBackend::new(88, 24)).unwrap();
-        term.draw(|f| {
-            render_hitl(f, &fat_request(), None, true, 0);
-        })
-        .unwrap();
-        let dump = term.backend().to_string();
-        assert!(
-            dump.contains("approve"),
-            "the operator cannot answer a gate whose keys are off-screen:\n{dump}"
-        );
-        assert!(dump.contains("deny"), "{dump}");
-    }
-
-    /// #939 §1: the command body must never be cut horizontally. A marker at
-    /// the very end of a long single-line command is the thing a destructive
-    /// suffix would occupy, so it is what the test looks for.
-    #[test]
-    fn a_long_command_is_wrapped_not_truncated() {
-        let req = HitlRequest {
-            tool_input: serde_json::json!({
-                "command": format!("git status {} && rm -rf /tmp/DANGER_MARKER", "-".repeat(120)),
-            }),
-            ..fat_request()
-        };
-        let mut term = Terminal::new(TestBackend::new(100, 40)).unwrap();
-        term.draw(|f| {
-            render_hitl(f, &req, None, true, 0);
-        })
-        .unwrap();
-        let dump = term.backend().to_string().replace(['\n', ' '], "");
-        assert!(
-            dump.contains("DANGER_MARKER"),
-            "the tail of the command was dropped — that is the whole defect:\n{}",
-            term.backend()
-        );
-    }
-
-    /// #939 §2: the notice counts HIDDEN rows. The old code printed the number
-    /// kept, so a taller box reported MORE hidden. Growing the terminal must
-    /// make the number go down, never up.
-    #[test]
-    fn hidden_line_count_falls_as_the_box_grows() {
-        let req = HitlRequest {
-            tool_input: serde_json::json!({ "command": "echo hi\n".repeat(400) }),
-            ..fat_request()
-        };
-        let hidden_at = |h: u16| -> usize {
-            let mut term = Terminal::new(TestBackend::new(100, h)).unwrap();
-            term.draw(|f| {
-                render_hitl(f, &req, None, true, 0);
-            })
-            .unwrap();
-            let dump = term.backend().to_string();
-            let tail = dump.split("… ").nth(1).expect("a residue notice");
-            tail.split_whitespace()
-                .next()
-                .and_then(|n| n.parse::<usize>().ok())
-                .expect("a numeric hidden count")
-        };
-        let small = hidden_at(20);
-        let large = hidden_at(40);
-        assert!(
-            large < small,
-            "a bigger box hid {large} lines vs {small} in a smaller one — \
-             the count is tracking box height, not residual content"
-        );
-    }
-
-    /// A page must never step over a row the operator was not shown. With a
-    /// fixed step of 5 and a box short enough to show 2 body rows, one PgDn
-    /// moved from rows 1-2 to rows 6-7 and rows 3-5 were never displayed —
-    /// silently, in the modal whose entire purpose is reading before running.
-    #[test]
-    fn paging_never_steps_over_an_unread_row() {
-        for visible in 1u16..=40 {
-            let step = super::hitl_scroll_step(visible);
-            assert!(
-                step <= visible,
-                "visible={visible} step={step}: a step wider than the window \
-                 skips rows that were never on screen"
-            );
-            assert!(step >= 1, "visible={visible}: a zero step cannot scroll");
-        }
-        // Before the first draw there is no measured height to clamp to.
-        assert_eq!(super::hitl_scroll_step(0), super::HITL_SCROLL_PAGE);
-    }
-
-    /// #939 §1+§3: scrolling reaches content that is off-screen at rest, and an
-    /// over-large offset is clamped rather than scrolling into blank space.
-    #[test]
-    fn paging_reveals_the_tail_and_clamps_at_the_end() {
-        let req = HitlRequest {
-            tool_input: serde_json::json!({ "command": (0..40).map(|i| format!("step{i}")).collect::<Vec<_>>().join("\n") }),
-            ..fat_request()
-        };
-        let draw = |scroll: u16| {
-            let mut term = Terminal::new(TestBackend::new(100, 20)).unwrap();
-            term.draw(|f| {
-                render_hitl(f, &req, None, true, scroll);
-            })
-            .unwrap();
-            term.backend().to_string()
-        };
-        assert!(
-            !draw(0).contains("step39"),
-            "sanity: the tail starts hidden"
-        );
-        assert!(
-            draw(200).contains("step39"),
-            "an offset past the end must clamp to the last page, not blank the body"
-        );
-    }
-
-    /// #939 §3: while the composer holds text the decision keys type instead of
-    /// deciding, so the modal must say so rather than advertising live keys.
-    #[test]
-    fn a_nonempty_composer_is_announced_on_the_key_row() {
-        let mut term = Terminal::new(TestBackend::new(100, 24)).unwrap();
-        term.draw(|f| {
-            render_hitl(f, &fat_request(), None, false, 0);
-        })
-        .unwrap();
-        let dump = term.backend().to_string().replace('\n', " ");
-        assert!(
-            dump.contains("Ctrl+U"),
-            "the operator gets no hint that y/a/A/n are inert:\n{dump}"
-        );
-    }
-}
-
-#[cfg(test)]
-mod chooser_floor_tests {
-    use super::super::app::App;
-    use super::super::complete::{Candidate, CompletionState};
-    use super::chooser_band_height;
-
-    fn option(display: &str, desc: &str) -> Candidate {
-        Candidate {
-            display: display.into(),
-            insert: display.into(),
-            desc: desc.into(),
-            has_children: false,
-        }
-    }
-
-    /// Three suggested replies, each with a description — the shape from the
-    /// report.
-    fn app_with_three_options() -> App {
-        let mut a = App::test_fixture();
-        a.completion = Some(CompletionState {
-            items: vec![
-                option("open the PR", "wait for CI, then tag"),
-                option("stronger model", "the 4B one fakes tool calls"),
-                option("leave it", "change nothing"),
-            ],
-            selected: 0,
-            spaced: true,
-        });
-        a
-    }
-
-    #[test]
-    fn the_chooser_leaves_the_transcript_more_than_three_rows() {
-        // Inline viewport is 20 rows; composer 3 + status 1 leaves 16.
-        let h = chooser_band_height(&app_with_three_options(), 20, 3);
-        assert!(
-            h <= 8,
-            "chooser took {h} rows, leaving the reply a peephole"
-        );
-    }
-
-    #[test]
-    fn a_short_terminal_still_gets_a_usable_chooser() {
-        // The floor must yield rather than squeeze the chooser out: it is the
-        // thing the operator has to act on.
-        let h = chooser_band_height(&app_with_three_options(), 12, 3);
-        assert!(h >= 5, "chooser unusable at {h} rows");
-    }
-
-    #[test]
-    fn ctrl_up_still_reaches_the_spaced_form() {
-        let mut a = app_with_three_options();
-        a.chooser_grow = 6;
-        assert_eq!(chooser_band_height(&a, 20, 3), 11);
-    }
-}
-
-#[cfg(test)]
-mod scroll_marker_tests {
-    use super::scroll_marker;
-
-    #[test]
-    fn silence_when_everything_fits() {
-        assert_eq!(scroll_marker(0, 0), None);
-    }
-
-    #[test]
-    fn following_the_tail_points_up() {
-        assert_eq!(scroll_marker(25, 0).as_deref(), Some(" ↑ 25 more · PgUp "));
-    }
-
-    #[test]
-    fn scrolled_back_points_the_way_home() {
-        assert_eq!(
-            scroll_marker(25, 10).as_deref(),
-            Some(" ↑ 15 · PgDn to follow ")
-        );
-    }
-}
-
-/// #940: the observability footer must never render a figure whose blank case
-/// is indistinguishable from a real one.
-#[cfg(test)]
-mod footer_cost_tests {
-    use super::super::footer::Pricing;
-    use super::footer_segments;
-
-    /// A model with no per-token price used to render an em dash in the exact
-    /// slot where a priced model renders `$0.125 est` — so "we cannot price
-    /// this" and "this cost nothing" were the same pixel. Two agents side by
-    /// side in a multiplexer is where it bit.
-    #[test]
-    fn an_unpriced_model_says_so_rather_than_going_blank() {
-        let unpriced = Pricing {
-            in_per_1k: None,
-            out_per_1k: None,
-            window: Some(32_000),
-        };
-        let out = footer_segments(10, 20, 100, 200, 3_762, &unpriced, None, 200);
-        assert!(
-            out.contains("no price"),
-            "an unpriced model must name the reason, got: {out}"
-        );
-        assert!(
-            !out.contains('\u{2014}'),
-            "the bare em dash reads as a cost of zero, got: {out}"
-        );
-    }
-
-    /// Control: pricing present still produces a number, so the fix above did
-    /// not simply replace the cost segment.
-    #[test]
-    fn a_priced_model_still_shows_the_estimate() {
-        let priced = Pricing {
-            in_per_1k: Some(3.0),
-            out_per_1k: Some(15.0),
-            window: Some(32_000),
-        };
-        let out = footer_segments(10, 20, 1_000, 1_000, 3_762, &priced, None, 200);
-        assert!(out.contains('$'), "priced model must show a cost: {out}");
-        assert!(!out.contains("no price"), "got: {out}");
-    }
-}
-
-/// #940: the status bar rendered a persisted channel-lifecycle word next to the
-/// live turn state, so it routinely read `⏵ 019ff831:working   ready`.
-#[cfg(test)]
-mod status_chip_tests {
-    use super::super::app::App;
-    use super::super::persist::ChannelMeta;
-    use super::render_status;
-    use ratatui::Terminal;
-    use ratatui::backend::TestBackend;
-
-    #[test]
-    fn the_channel_chip_states_the_id_and_leaves_the_state_word_to_the_live_one() {
-        let mut app = App::test_fixture();
-        app.channel = Some(ChannelMeta {
-            id: "019ff831dead".into(),
-        });
-        let mut term = Terminal::new(TestBackend::new(120, 1)).unwrap();
-        term.draw(|f| render_status(f, &app, f.area())).unwrap();
-        let dump = term.backend().to_string();
-
-        assert!(dump.contains("019ff831"), "chip must show the id: {dump}");
-        assert!(dump.contains("ready"), "live turn state stays: {dump}");
-        assert!(
-            !dump.contains("019ff831:"),
-            "the chip must not append a second state word: {dump}"
-        );
-    }
-}
-
-#[cfg(test)]
-mod gap_tests {
-    use super::super::app::{ChatMsg, Role};
-    use super::super::step::StepCard;
-    use super::super::theme::{DARK, MUR};
-    use super::{gap_row, wants_gap_before};
-
-    fn card_msg() -> ChatMsg {
-        ChatMsg::tool_for_test(StepCard::new(
-            "s".into(),
-            "bash".into(),
-            serde_json::json!({"command": "ls"}),
-        ))
-    }
-
-    /// The whole point: a run of tool calls is one block of work, not N
-    /// separate events. Ten calls used to cost ten gap rows.
-    #[test]
-    fn a_tool_call_does_not_open_a_gap() {
-        assert!(!wants_gap_before(&card_msg()));
-    }
-
-    /// Control — if this ever flips, gaps stop marking anything at all.
-    #[test]
-    fn a_spoken_turn_still_opens_a_gap() {
-        assert!(wants_gap_before(&ChatMsg::for_test(Role::User, "hi")));
-        assert!(wants_gap_before(&ChatMsg::for_test(Role::Agent, "hello")));
-        assert!(wants_gap_before(&ChatMsg::for_test(Role::System, "note")));
-    }
-
-    /// One builder for the row, so the two emit paths cannot drift into
-    /// different-looking gaps.
-    #[test]
-    fn the_gap_row_follows_the_skin() {
-        let (u, a) = (
-            ChatMsg::for_test(Role::User, "hi"),
-            ChatMsg::for_test(Role::Agent, "hello"),
-        );
-        let blank: String = gap_row(&DARK, Some(&u), &a)
-            .spans
-            .iter()
-            .map(|s| s.content.as_ref())
-            .collect();
-        assert!(
-            blank.trim().is_empty(),
-            "dark skin uses a blank row: {blank:?}"
-        );
-
-        let ruled: String = gap_row(&MUR, Some(&u), &a)
-            .spans
-            .iter()
-            .map(|s| s.content.as_ref())
-            .collect();
-        if MUR.show_separator {
-            assert!(ruled.contains('─'), "a ruled skin draws a rule: {ruled:?}");
-        }
-    }
-}
-
-#[cfg(test)]
-mod block_tests {
-    use super::super::app::{App, ChatMsg, Role};
-    use super::super::step::StepCard;
-    use super::message_block;
-
-    fn card() -> ChatMsg {
-        ChatMsg::tool_for_test(StepCard::new(
-            "s".into(),
-            "bash".into(),
-            serde_json::json!({"command": "ls"}),
-        ))
-    }
-
-    fn text(lines: &[ratatui::text::Line<'static>]) -> Vec<String> {
-        lines
-            .iter()
-            .map(|l| l.spans.iter().map(|s| s.content.as_ref()).collect())
-            .collect()
-    }
-
-    fn is_gap(l: &str) -> bool {
-        l.trim().is_empty() || l.chars().all(|c| c == '─')
-    }
-
-    /// The bug: the viewport and the emit path disagreed. Same message, same
-    /// index — the rows must be identical, gap included.
-    #[test]
-    fn measured_and_rendered_blocks_agree() {
-        let app = App::test_fixture();
-        let m = ChatMsg::for_test(Role::Agent, "hello");
-        let rendered = text(&message_block(&app, 3, &m, 0, false));
-        let measured = text(&message_block(&app, 3, &m, 0, true));
-        assert_eq!(
-            rendered.iter().filter(|l| is_gap(l)).count(),
-            measured.iter().filter(|l| is_gap(l)).count(),
-            "the row the flush decision counts must be the row that gets drawn"
-        );
-    }
-
-    /// A tool call is a step inside the turn before it.
-    #[test]
-    fn a_step_card_opens_no_gap() {
-        let app = App::test_fixture();
-        let lines = text(&message_block(&app, 3, &card(), 0, false));
-        assert!(!lines.first().is_some_and(|l| is_gap(l)), "{lines:?}");
-    }
-
-    /// Control — if this flips, gaps stop marking anything.
-    #[test]
-    fn a_spoken_turn_opens_a_gap() {
-        let app = App::test_fixture();
-        let m = ChatMsg::for_test(Role::User, "hi");
-        let lines = text(&message_block(&app, 3, &m, 0, false));
-        assert!(lines.first().is_some_and(|l| is_gap(l)), "{lines:?}");
-    }
-
-    /// Nothing opens the transcript with a blank row, and a continuation
-    /// resumes a message whose head is already committed above.
-    #[test]
-    fn the_first_message_and_continuations_open_no_gap() {
-        let app = App::test_fixture();
-        let m = ChatMsg::for_test(Role::User, "hi");
-        let first = text(&message_block(&app, 0, &m, 0, false));
-        assert!(!first.first().is_some_and(|l| is_gap(l)), "{first:?}");
-        let resumed = text(&message_block(&app, 3, &m, 2, false));
-        assert!(!resumed.first().is_some_and(|l| is_gap(l)), "{resumed:?}");
-    }
-}
-
-#[cfg(test)]
-mod transcript_chrome_tests {
-    use super::super::app::App;
-    use super::render_transcript;
-    use ratatui::Terminal;
-    use ratatui::backend::TestBackend;
-
-    fn painted() -> String {
-        let mut app = App::test_fixture();
-        let mut term = Terminal::new(TestBackend::new(60, 10)).unwrap();
-        term.draw(|f| {
-            let area = f.area();
-            render_transcript(f, &mut app, area);
-        })
-        .unwrap();
-        term.backend()
-            .buffer()
-            .content()
-            .iter()
-            .map(|c| c.symbol())
-            .collect::<Vec<_>>()
-            .chunks(60)
-            .map(|row| row.concat())
-            .collect::<Vec<_>>()
-            .join("\n")
-    }
-
-    /// The agent name is already in the status-bar badge. A full-width rule
-    /// spent restating it is the clearest kind of cell that has not earned its
-    /// place.
-    #[test]
-    fn the_transcript_border_carries_no_title() {
-        let out = painted();
-        assert!(!out.contains("chat ·"), "title is back:\n{out}");
-    }
-
-    /// No rule at all. The composer's own titled top border marks the seam
-    /// beneath this block, and the top rule only ever hosted the scroll
-    /// marker — one line too many on a screen already ruled between speakers.
-    #[test]
-    fn no_edge_is_drawn() {
-        let out = painted();
-        let ruled: Vec<usize> = out
-            .lines()
-            .enumerate()
-            .filter(|(_, l)| l.chars().filter(|c| *c == '─').count() > 10)
-            .map(|(i, _)| i)
-            .collect();
-        assert!(
-            ruled.is_empty(),
-            "expected no rule, got rows {ruled:?}\n{out}"
-        );
-    }
-}
-
-#[cfg(test)]
-mod welcome_surface_tests {
-    use super::super::app::{App, ChatMsg, Role};
-    use super::super::theme::MUR;
-    use super::super::welcome::MASCOT_REST;
-    use super::{message_block, render_transcript};
-    use ratatui::Terminal;
-    use ratatui::backend::TestBackend;
-    use ratatui::layout::Rect;
-
-    fn dump(app: &mut App) -> String {
-        let mut term = Terminal::new(TestBackend::new(100, 40)).unwrap();
-        term.draw(|f| render_transcript(f, app, Rect::new(0, 0, 100, 40)))
-            .unwrap();
-        term.backend().to_string()
-    }
-
-    /// A slash command's notice is not a conversation. Before, the first
-    /// `/skills` ended the welcome: the mascot vanished, the viewport shrank
-    /// to the chat height, and a wiped screen showed five rows of notice over
-    /// a blank slab. The welcome stays until someone actually speaks.
-    #[test]
-    fn a_slash_notice_keeps_the_welcome_on_screen() {
-        let mut app = App::test_fixture();
-        app.push_system("skin changed to mur");
-        let d = dump(&mut app);
-        assert!(
-            d.contains(MASCOT_REST[1]),
-            "mascot gone after a notice:\n{d}"
-        );
-        assert!(d.contains("skin changed to mur"), "notice missing:\n{d}");
-
-        // Control: the welcome leaves with the head of the band, not with
-        // the first spoken turn (see `band_growth_tests`).
-        app.messages.push(ChatMsg::for_test(Role::User, "hi"));
-        app.flushed_upto = 1;
-        let d = dump(&mut app);
-        assert!(
-            !d.contains(MASCOT_REST[1]),
-            "welcome outlived its flush:\n{d}"
-        );
-    }
-
-    /// Three `/skin` switches drew three rules under the light skin. A rule
-    /// marks a change of speaker; a run of notices is one speaker (the UI).
-    #[test]
-    fn consecutive_notices_draw_no_rule_and_turns_still_do() {
-        let mut app = App::test_fixture();
-        app.theme = &MUR;
-        const { assert!(MUR.show_separator, "test needs a ruled skin") };
-        app.push_system("skin changed to light");
-        app.push_system("skin changed to mur");
-        app.messages.push(ChatMsg::for_test(Role::User, "hi"));
-        app.messages.push(ChatMsg::for_test(Role::Agent, "hello"));
-        let text = |i: usize| -> String {
-            message_block(&app, i, &app.messages[i], 0, false)
-                .iter()
-                .flat_map(|l| l.spans.iter().map(|s| s.content.to_string()))
-                .collect()
-        };
-        assert!(
-            !text(1).contains('─'),
-            "notice after notice ruled: {:?}",
-            text(1)
-        );
-        assert!(
-            !text(2).contains('─'),
-            "user after notice ruled: {:?}",
-            text(2)
-        );
-        assert!(
-            text(3).contains('─'),
-            "agent after user not ruled: {:?}",
-            text(3)
-        );
-    }
-}
-
-/// The live band grows upward: what no longer fits is pushed into scrollback
-/// (directly above the band, still on screen), never hidden behind a marker.
-#[cfg(test)]
-mod band_growth_tests {
-    use super::super::app::{App, ChatMsg, RenderMode, Role};
-    use super::super::complete::{Candidate, CompletionState};
-    use super::super::welcome::MASCOT_REST;
-    use super::{flush_finished, render, render_transcript};
-    use ratatui::backend::TestBackend;
-    use ratatui::layout::Rect;
-    use ratatui::{Terminal, TerminalOptions, Viewport};
-
-    fn option(display: &str) -> Candidate {
-        Candidate {
-            display: display.into(),
-            insert: display.into(),
-            desc: String::new(),
-            has_children: false,
-        }
-    }
-
-    /// A finished exchange with three suggested replies pending — the shape
-    /// from the report: the reply the operator must read to choose.
-    fn app_with_open_chooser() -> App {
-        let mut app = App::test_fixture();
-        app.render_mode = RenderMode::Inline;
-        let long = (1..=8)
-            .map(|i| format!("earlier line {i}"))
-            .collect::<Vec<_>>()
-            .join("\n");
-        app.messages.push(ChatMsg::for_test(Role::Agent, &long));
-        app.messages
-            .push(ChatMsg::for_test(Role::User, "先建 Messaging API channel"));
-        app.messages.push(ChatMsg::for_test(
-            Role::Agent,
-            "Messaging API 建好以後跟我說",
-        ));
-        app.completion = Some(CompletionState {
-            items: vec![option("建好了"), option("卡住了"), option("接著建")],
-            selected: 0,
-            spaced: true,
-        });
-        app
-    }
-
-    /// The chooser takes rows from the band; before, the band answered by
-    /// hiding rows above ("↑ 7 more · PgUp") — the reply behind the chooser
-    /// read as lost. Now the overflow is flushed up into scrollback, where it
-    /// stays readable while the operator picks.
-    #[test]
-    fn an_open_chooser_pushes_the_reply_up_instead_of_hiding_it() {
-        let mut app = app_with_open_chooser();
-        let mut term = Terminal::with_options(
-            TestBackend::new(100, 60),
-            TerminalOptions {
-                viewport: Viewport::Inline(20),
-            },
-        )
-        .unwrap();
-        flush_finished(&mut term, &mut app, 20).unwrap();
-        term.draw(|f| render(f, &mut app)).unwrap();
-        let d = term.backend().to_string();
-        assert!(!d.contains("PgUp"), "rows hidden behind the chooser:\n{d}");
-        assert!(
-            d.contains("earlier line 1"),
-            "flushed reply not on screen:\n{d}"
-        );
-        assert!(
-            d.contains("建好以後跟我說"),
-            "latest reply not on screen:\n{d}"
-        );
-    }
-
-    /// The band's top rule was one more line on a screen full of them. It
-    /// only ever carried the scroll marker, which now paints on the first
-    /// row by itself when — and only when — rows are hidden.
-    #[test]
-    fn no_rule_above_the_band_and_the_marker_still_shows_when_needed() {
-        let mut app = App::test_fixture();
-        app.messages.push(ChatMsg::for_test(Role::User, "hi"));
-        app.messages.push(ChatMsg::for_test(Role::Agent, "hello"));
-        let mut term = Terminal::new(TestBackend::new(100, 30)).unwrap();
-        term.draw(|f| render_transcript(f, &mut app, Rect::new(0, 0, 100, 30)))
-            .unwrap();
-        let d = term.backend().to_string();
-        assert!(!d.contains('─'), "a rule above the band:\n{d}");
-        assert!(!d.contains("PgUp"), "marker with nothing hidden:\n{d}");
-
-        // Control: squeeze the same transcript so rows really are hidden.
-        app.flushed_upto = 0;
-        let mut term = Terminal::new(TestBackend::new(100, 3)).unwrap();
-        term.draw(|f| render_transcript(f, &mut app, Rect::new(0, 0, 100, 3)))
-            .unwrap();
-        let d = term.backend().to_string();
-        assert!(d.contains("more · PgUp"), "hidden rows unmarked:\n{d}");
-    }
-
-    /// The mascot is the head of the band, not a splash that the first turn
-    /// replaces: it stays until the band fills and the flush carries it up.
-    #[test]
-    fn the_mascot_stays_until_the_first_flush() {
-        let mut app = App::test_fixture();
-        app.messages.push(ChatMsg::for_test(Role::User, "hi"));
-        app.messages.push(ChatMsg::for_test(Role::Agent, "hello"));
-        let dump = |app: &mut App| {
-            let mut term = Terminal::new(TestBackend::new(100, 40)).unwrap();
-            term.draw(|f| render_transcript(f, app, Rect::new(0, 0, 100, 40)))
-                .unwrap();
-            term.backend().to_string()
-        };
-        let d = dump(&mut app);
-        assert!(d.contains(MASCOT_REST[1]), "mascot gone after a turn:\n{d}");
-        assert!(d.contains("hello"), "reply missing under the mascot:\n{d}");
-
-        // Control: once the head is in scrollback the band no longer paints it.
-        app.flushed_upto = 1;
-        let d = dump(&mut app);
-        assert!(!d.contains(MASCOT_REST[1]), "mascot painted twice:\n{d}");
-    }
 }
